+++ CHANGELOG.md
@@ -6,6 +6,7 @@ For change log format, use [Keep a Changelog] (http://keepachangelog.com/).
 ## [1.3.1] - 2020-12-09
 ### Changed
 - Escape dot at first character line, on SMTP dialog with quoted_printable_encode()
+- Thrown an exception if an attachment file does not exist
 
 ## [1.3] - 2020-11-05
 ### Added